Fixed lot of BUGs about hard delete of comment.

+++ app/helpers/comments_helper.rb
@@ -6,7 +6,7 @@ module CommentsHelper
   end
 
   def delete_button(comment)
-    link_to comment_path(comment), :class => "pull-right btn", :role => "button", :title => t("delete"), method: :delete do
+    link_to comment_path(comment), :class => "pull-right btn", :role => "button", :title => t("delete"), method: :delete, data: { confirm: 'Are you sure? Your idea will be deleted permanently with your vote.' } do
       content_tag(:span, '', :class => 'glyphicon glyphicon-remove dark-grey', 'aria-hidden' => 'true')
     end
   end
